Migration to basics app done
