--- I am rich/ViewController.swift
@@ -1,14 +1,0 @@
-
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-     
-    }
-
-
-}
-
