+++ iosbasic.xcodeproj/xcuserdata/vmanikkam.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -4,7 +4,7 @@
 <dict>
 	<key>SchemeUserState</key>
 	<dict>
-		<key>I am rich.xcscheme_^#shared#^_</key>
+		<key>iosbasic.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
 			<integer>0</integer>
