+++ iosbasic/ViewController.swift
@@ -1,0 +1,19 @@
+//
+//  ViewController.swift
+//  iosbasic
+//
+//  Created by Venkatasalam Manikkam on 01/10/24.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+    }
+
+
+}
+
